ipcollect-microservice/requirements.txt
jsonschema==3.2.0
ncclient==0.6.13
xmltodict==0.12.0
pydot==1.4.2
python-qpid-proton==0.38.0
jwt==1.3.1
